+++ src/gui/CustomDeltaTEquationDialog.hpp
@@ -59,9 +59,9 @@ private:
 	QSettings* conf;
 	StelCore* core;
 
-	float year;
-	float ndot;
-	Vec3f coeff;
+	double year;
+	double ndot;
+	Vec3d coeff;
 
 	void setDescription(void) const;
 };
